Aktualizacja JS

+++ ProjektRestauracja/DataStore/ProductRepository.js
@@ -36,7 +36,7 @@ var __generator = (this && this.__generator) || function (thisArg, body) {
     }
 };
 exports.__esModule = true;
-exports.ProductDemandList = exports.ProductRepository = void 0;
+exports.ProductRepository = void 0;
 var mongoose_1 = require("mongoose");
 var ProductRepository = /** @class */ (function () {
     function ProductRepository() {
@@ -57,39 +57,57 @@ var ProductRepository = /** @class */ (function () {
                         _a.sent();
                         products = [
                             {
-                                name: 'Cola',
+                                name: "Coca_Cola_Can",
                                 price: 2.5,
-                                quantity: 10
-                            }, {
-                                name: 'Fanta',
-                                price: 3.5,
-                                quantity: 10
-                            }, {
-                                name: 'Sprite',
-                                price: 4.5,
-                                quantity: 10
-                            }, {
-                                name: 'Coca-Cola Zero',
-                                price: 1.5,
-                                quantity: 10
+                                quantity: 250
                             },
                             {
-                                name: 'Red_Wine_Bottle',
+                                name: "Fanta_Can",
+                                price: 2.5,
+                                quantity: 250
+                            },
+                            {
+                                name: "Carrot",
+                                price: 1.5,
+                                quantity: 100
+                            },
+                            {
+                                name: "Parsley",
+                                price: 1.5,
+                                quantity: 100
+                            },
+                            {
+                                name: "Onion",
+                                price: 1.5,
+                                quantity: 100
+                            },
+                            {
+                                name: "Tomato",
+                                price: 1.5,
+                                quantity: 100
+                            },
+                            {
+                                name: "Cucumber",
+                                price: 1.5,
+                                quantity: 100
+                            },
+                            {
+                                name: "Red_Wine_Bottle",
                                 price: 5,
                                 quantity: 50
                             },
                             {
-                                name: 'Chicken',
+                                name: "Chicken",
                                 price: 3,
                                 quantity: 100
                             },
                             {
-                                name: 'Mushroom',
+                                name: "Mushroom",
                                 price: 1,
                                 quantity: 200
                             },
                             {
-                                name: 'Cabbage',
+                                name: "Cabbage",
                                 price: 2,
                                 quantity: 500
                             }
@@ -114,20 +132,33 @@ var ProductRepository = /** @class */ (function () {
     };
     ProductRepository.prototype.addProduct = function (product) {
         return __awaiter(this, void 0, void 0, function () {
+            var alreadyExists, existsAfter;
             return __generator(this, function (_a) {
                 switch (_a.label) {
                     case 0: return [4 /*yield*/, (0, mongoose_1.connect)('mongodb+srv://Admin:<AdminAdmin>@cluster0.tpgqv.mongodb.net/?retryWrites=true&w=majority')];
                     case 1:
                         _a.sent();
+                        return [4 /*yield*/, this.ProductModel.exists({ name: product.name })];
+                    case 2:
+                        alreadyExists = _a.sent();
+                        if (alreadyExists)
+                            return [2 /*return*/, false];
                         return [4 /*yield*/, this.ProductModel
                                 .create(product)
                                 .then(function () {
-                                console.log("Product" + product.productId + "has been added");
+                                console.log("Product " + product.name + " has been added!");
                             })["catch"](function (err) {
                                 console.log(err);
                             })];
-                    case 2:
+                    case 3:
                         _a.sent();
+                        return [4 /*yield*/, this.ProductModel.findOne({ name: product.name })];
+                    case 4:
+                        existsAfter = _a.sent();
+                        if (existsAfter)
+                            return [2 /*return*/, true];
+                        else
+                            return [2 /*return*/, false];
                         return [2 /*return*/];
                 }
             });
@@ -135,20 +166,33 @@ var ProductRepository = /** @class */ (function () {
     };
     ProductRepository.prototype.deleteProductByName = function (productName) {
         return __awaiter(this, void 0, void 0, function () {
+            var exists, existsAfter;
             return __generator(this, function (_a) {
                 switch (_a.label) {
                     case 0: return [4 /*yield*/, (0, mongoose_1.connect)('mongodb+srv://Admin:<AdminAdmin>@cluster0.tpgqv.mongodb.net/?retryWrites=true&w=majority')];
                     case 1:
                         _a.sent();
+                        return [4 /*yield*/, this.ProductModel.exists({ name: productName })];
+                    case 2:
+                        exists = _a.sent();
+                        if (!exists)
+                            return [2 /*return*/, false];
                         return [4 /*yield*/, this.ProductModel
                                 .deleteOne({ name: productName })
                                 .then(function () {
-                                console.log("Product" + productName + " has been deleted!");
+                                console.log("Product " + productName + " has been deleted!");
                             })["catch"](function (err) {
                                 console.log(err);
                             })];
-                    case 2:
+                    case 3:
                         _a.sent();
+                        return [4 /*yield*/, this.ProductModel.findOne({ name: productName })];
+                    case 4:
+                        existsAfter = _a.sent();
+                        if (!existsAfter)
+                            return [2 /*return*/, true];
+                        else
+                            return [2 /*return*/, false];
                         return [2 /*return*/];
                 }
             });
@@ -165,12 +209,10 @@ var ProductRepository = /** @class */ (function () {
                         return [4 /*yield*/, this.ProductModel.findOne({ name: productName })];
                     case 2:
                         product = _a.sent();
-                        if (product) {
+                        if (product)
                             return [2 /*return*/, product];
-                        }
-                        else {
-                            return [2 /*return*/, null];
-                        }
+                        else
+                            return [2 /*return*/, false];
                         return [2 /*return*/];
                 }
             });
@@ -178,17 +220,25 @@ var ProductRepository = /** @class */ (function () {
     };
     ProductRepository.prototype.getProduct = function () {
         return __awaiter(this, void 0, void 0, function () {
+            var products;
             return __generator(this, function (_a) {
                 switch (_a.label) {
                     case 0: return [4 /*yield*/, (0, mongoose_1.connect)('mongodb+srv://Admin:<AdminAdmin>@cluster0.tpgqv.mongodb.net/?retryWrites=true&w=majority')];
                     case 1:
                         _a.sent();
-                        return [2 /*return*/, this.ProductModel.find()];
+                        return [4 /*yield*/, this.ProductModel.find({})];
+                    case 2:
+                        products = _a.sent();
+                        if (products.length > 0)
+                            return [2 /*return*/, products];
+                        else
+                            return [2 /*return*/, false];
+                        return [2 /*return*/];
                 }
             });
         });
     };
-    ProductRepository.prototype.updateProduct = function (productName, product) {
+    ProductRepository.prototype.updateProductByName = function (productName, product) {
         return __awaiter(this, void 0, void 0, function () {
             var productToUpdate;
             return __generator(this, function (_a) {
@@ -214,11 +264,8 @@ var ProductRepository = /** @class */ (function () {
                             })];
                     case 3:
                         _a.sent();
-                        return [3 /*break*/, 5];
-                    case 4:
-                        console.log("Product " + productName + " does not exist!");
-                        _a.label = 5;
-                    case 5: return [2 /*return*/];
+                        return [2 /*return*/, true];
+                    case 4: return [2 /*return*/, false];
                 }
             });
         });
@@ -226,36 +273,35 @@ var ProductRepository = /** @class */ (function () {
     return ProductRepository;
 }());
 exports.ProductRepository = ProductRepository;
-var ProductDemandList = /** @class */ (function () {
-    function ProductDemandList() {
-        this.productNames = [];
-        this.productQuantities = [];
-        this.productNames = [];
-        this.productQuantities = [];
-    }
-    ProductDemandList.prototype.AddProduct = function (product) {
-        var index = this.productNames.indexOf(product.name);
-        if (index == -1) {
-            this.productNames.push(product.name);
-            this.productQuantities.push(product.quantity);
-        }
-        else {
-            this.productQuantities[index] += product.quantity;
-        }
-    };
-    ProductDemandList.prototype.GetProductName = function () {
-        return this.productNames;
-    };
-    ProductDemandList.prototype.GetDemandList = function () {
-        return this;
-    };
-    ProductDemandList.prototype.GetProductQuantityByName = function (name) {
-        var index = this.productNames.indexOf(name);
-        return this.productQuantities[index];
-    };
-    ProductDemandList.prototype.GetProductQuantityByIndex = function (index) {
-        return this.productQuantities[index];
-    };
-    return ProductDemandList;
-}());
-exports.ProductDemandList = ProductDemandList;
+// export class ProductDemandList
+// {
+//     productNames: string[]=[];
+//     productQuantities: number[]=[];
+//     constructor(){
+//         this.productNames = [];
+//         this.productQuantities = [];
+//     }
+//     AddProduct(product: Product){
+//         let index = this.productNames.indexOf(product.name);
+//         if(index == -1){
+//             this.productNames.push(product.name);
+//             this.productQuantities.push(product.quantity);
+//         }else{
+//             this.productQuantities[index] += product.quantity;
+//         }
+//     }
+//     GetProductName()
+//     {
+//         return this.productNames;
+//     }
+//     GetDemandList(){
+//         return this;
+//     }
+//     GetProductQuantityByName(name: string){
+//         let index = this.productNames.indexOf(name);
+//         return this.productQuantities[index];
+//     }
+//     GetProductQuantityByIndex(index: number){
+//         return this.productQuantities[index];
+//     }
+// }
